+++ src/controls/List/Component/index.js
@@ -19,6 +19,8 @@ export default class LayoutComponent extends Component {
     onChange: PropTypes.func,
     currentState: PropTypes.object,
     translations: PropTypes.object,
+    indentDisabled: PropTypes.bool,
+    outdentDisabled: PropTypes.bool,
   };
 
   options: Array = ['unordered', 'ordered', 'indent', 'outdent'];
@@ -41,7 +43,13 @@ export default class LayoutComponent extends Component {
   // todo: evaluate refactoring this code to put a loop there and in other places also in code
   // hint: it will require moving click handlers
   renderInFlatList(): Object {
-    const { config, currentState: { listType }, translations } = this.props;
+    const {
+      config,
+      currentState: { listType },
+      translations,
+      indentDisabled,
+      outdentDisabled
+    } = this.props;
     const { options, unordered, ordered, indent, outdent, className } = config;
     return (
       <div className={classNames('rdw-list-wrapper', className)} aria-label="rdw-list-control">
@@ -71,6 +79,7 @@ export default class LayoutComponent extends Component {
         </Option>}
         {options.indexOf('indent') >= 0 && <Option
           onClick={this.indent}
+          disabled={indentDisabled}
           className={classNames(indent.className)}
           title={indent.title || translations['components.controls.list.indent']}
         >
@@ -81,6 +90,7 @@ export default class LayoutComponent extends Component {
         </Option>}
         {options.indexOf('outdent') >= 0 && <Option
           onClick={this.outdent}
+          disabled={outdentDisabled}
           className={classNames(outdent.className)}
           title={outdent.title || translations['components.controls.list.outdent']}
         >
@@ -126,6 +136,7 @@ export default class LayoutComponent extends Component {
           .map((option, index) => (<DropdownOption
             key={index}
             value={option}
+            disabled={this.props[`${option}Disabled`]}
             className={classNames('rdw-list-dropdownOption', config[option].className)}
             active={listType === option}
             title={config[option].title || translations[`components.controls.list.${option}`]}