+++ stories/Basic/index.js
@@ -8,6 +8,7 @@ const Basic = () => (<div className="rdw-storybook-root">
     toolbarClassName="rdw-storybook-toolbar"
     wrapperClassName="rdw-storybook-wrapper"
     editorClassName="rdw-storybook-editor"
+    handlePastedText={() => { console.log('testing...!') }}
   />
 </div>);